requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
authlib==1.6.4
backoff==2.2.1
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
comm==0.2.3
cryptography==46.0.1
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
executing==2.2.1
frozenlist==1.7.0
gitdb==4.0.12
gitingest==0.3.1
gitpython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
ipykernel==6.30.1
ipython==9.5.0
ipython-pygments-lexers==1.1.1
isort==6.0.1
jedi==0.19.2
jinja2==3.1.6
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-pydantic==0.6
jsonschema-specifications==2025.9.1
jupyter-client==8.6.3
jupyter-core==5.8.1
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-mcp-adapters==0.1.10
langchain-mongodb==0.7.0
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.8
langgraph-store-mongodb==0.1.0
langsmith==0.4.28
lark==1.2.2
loguru==0.7.3
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.14.1
-e file:///Users/raghib/Desktop/multi-agent/mcp-use/mcp-use
multidict==6.6.4
mypy-extensions==1.1.0
narwhals==2.5.0
nest-asyncio==1.6.0
numpy==2.3.3
openai==1.108.0
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.2
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
posthog==6.7.5
prompt-toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==21.0.0
pycparser==2.23
pydantic==2.11.9
pydantic-core==2.33.2
pydantic-settings==2.10.1
pydeck==0.9.1
pygments==2.19.2
pymongo==4.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.1.0
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
ruff==0.13.1
scarf-sdk==0.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.43
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.48.0
streamlit==1.49.1
tenacity==9.1.2
tiktoken==0.11.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.36.0
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.25.0
